LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/rendering/the-details-element/details-focus-steps-expected.txt

First
Finally Third

FAIL Focus flows into second when details is open assert_equals: expected Element node <a id="second" href="#">Second</a> but got Element node <summary id="first">First</summary>
FAIL Focus flows into third when details is closed assert_equals: expected Element node <button id="third">Third</button> but got Element node <summary id="first">First</summary>
